requirements.txt
openai==1.61.0  # OpenAI API client
python-dotenv
flask                  # For building a simple web UI
faiss-cpu              # For vector-based memory search
azure-search-documents # If using Azure Cognitive Search for memory
semantic-kernel==1.26.1